Add full stops to the end of hint texts
At the moment we're inconsistent with the use of a full stop at the end
of hint texts. We could either remove them all and rely on the frontend
to add them, or put them all in here explicitly. We've decided to go
with the latter to avoid unnecessary extra logic in the frontend and
make it more flexible.