Add configuration for yamllint
This configuration is based on the default configuration for yamllint
with only a few changes.

I've set the line length limit to 88. This is based on my experience of
using Black which has the same line length limit. There's an explanation
of this here: https://black.readthedocs.io/en/stable/the_black_code_style/current_style.html#line-length

Because of how we load the YAML files, we also don't need to declare
each one being a document with the `---` mark.